Add /services/products.js
