--- src/services/products/generateAddedProductPayload.js
@@ -1,11 +1,0 @@
-const generateAddedProductPayload = ({ product, selectedQuantity }) => {
-  const numberSystem = 10;
-
-  return {
-    id: product.id,
-    quantity: parseInt(selectedQuantity, numberSystem),
-    attributes: product
-  };
-};
-
-export default generateAddedProductPayload;
